+++ Object.h
@@ -19,51 +19,51 @@ public:
 
     vector<SDL_Texture*> getTextures() { return textures; }
     SDL_Texture* getTexture(int index);
-    double getX1();
-    double getX2();
-    double getY1();
-    double getY2();
-    double getX1HitBox();
-    double getX2HitBox();
-    double getY1HitBox();
-    double getY2HitBox(); 
+    float getX1();
+    float getX2();
+    float getY1();
+    float getY2();
+    float getX1HitBox();
+    float getX2HitBox();
+    float getY1HitBox();
+    float getY2HitBox(); 
     int getWIDTH();
     int getHEIGHT();
     int getSpeed();
-    double getSpeedAnimations();
+    float getSpeedAnimations();
     virtual bool isDead();
     bool endDeadAnimation();
 
     void setSprite(vector<SDL_Texture*> textures);
     void setSprite(SDL_Texture* texture, int index);
-    void setX(double X);
-    void setY(double Y);
-    void setSpeed(double speed);
-    void setSpeedAnimations(double speedAnimations);
+    void setX(float X);
+    void setY(float Y);
+    void setSpeed(float speed);
+    void setSpeedAnimations(float speedAnimations);
     void setDead();
     
     
-    virtual void animationBase(double deltaTime) = 0;
-    virtual void animationDead(double deltaTime) = 0;
+    virtual void animationBase(float deltaTime) = 0;
+    virtual void animationDead(float deltaTime) = 0;
     virtual void draw(SDL_Renderer* renderer);
-    virtual void update(double deltaTime) = 0;
+    virtual void update(float deltaTime) = 0;
 
 protected:
     std::vector<SDL_Texture*> textures;
-    double X1;
-    double X2;
-    double Y1;
-    double Y2;
-    double hitBoxMultiplication;
-    double X1HitBox;
-    double X2HitBox;
-    double Y1HitBox;
-    double Y2HitBox; 
+    float X1;
+    float X2;
+    float Y1;
+    float Y2;
+    float hitBoxMultiplication;
+    float X1HitBox;
+    float X2HitBox;
+    float Y1HitBox;
+    float Y2HitBox; 
     int WIDTH;
     int HEIGHT;
-    double deltaTime;
-    double speed;
-    double speedAnimations;
+    float deltaTime;
+    float speed;
+    float speedAnimations;
     int indexTexture;
     bool dead;
     bool deadAnimationEnd;
